Handling Postgres Bucket Connection errors

+++ buelon/cython/c_bucket.pyx
@@ -13,6 +13,7 @@ import time
 
 import psutil
 import redis
+import psycopg2
 
 import buelon.helpers.postgres
 
@@ -24,6 +25,8 @@ except ModuleNotFoundError:
 
 
 # Environment variables for client and server configuration
+USING_POSTGRES: bool = os.environ.get('USING_POSTGRES_BUCKET', 'false') == 'true'
+
 REDIS_HOST: str = os.environ.get('REDIS_HOST', 'null')
 REDIS_PORT: int = int(os.environ.get('REDIS_PORT', 6379))
 REDIS_DB: int = int(os.environ.get('REDIS_DB', 0))
@@ -32,7 +35,7 @@ try:
     REDIS_EXPIRATION = int(REDIS_EXPIRATION)
 except ValueError:
     REDIS_EXPIRATION = None
-USING_REDIS: bool = False  # REDIS_HOST != 'null'
+USING_REDIS: bool = os.environ.get('USING_REDIS', 'false') == 'true'  # False  # REDIS_HOST != 'null'
 
 BUCKET_CLIENT_HOST: str = os.environ.get('BUCKET_CLIENT_HOST', 'localhost')
 BUCKET_CLIENT_PORT: int = int(os.environ.get('BUCKET_CLIENT_PORT', 61535))
@@ -95,14 +98,17 @@ def retry_connection(func: callable):
     Returns:
         callable: The decorated function.
     """
-    def wrapper(*args, **kwargs):
+    def wrapper(self, *args, **kwargs):
         tries = 4
         kwargs['timeout'] = kwargs.get('timeout', 60 * 5.)
         for i in range(tries):
             try:
-                return func(*args, **kwargs)
+                return func(self, *args, **kwargs)
             except (TimeoutError, ConnectionResetError):
                 kwargs['timeout'] *= 2
+            except psycopg2.OperationalError:
+                time.sleep((i + 1) * 5.)
+                self.db = buelon.helpers.postgres.get_postgres_from_env()
         raise
     return wrapper
 
@@ -134,13 +140,14 @@ class Client:
         self.PORT = BUCKET_CLIENT_PORT
         self.HOST = BUCKET_CLIENT_HOST
 
-        self.db = buelon.helpers.postgres.get_postgres_from_env()
-        with self.db.connect() as conn:
-            cur = conn.cursor()
-            cur.execute('CREATE TABLE IF NOT EXISTS bucket (key TEXT PRIMARY KEY, data BYTEA, epoch REAL);')
-            conn.commit()
+        if USING_POSTGRES:
+            self.db = buelon.helpers.postgres.get_postgres_from_env()
+            with self.db.connect() as conn:
+                cur = conn.cursor()
+                cur.execute('CREATE TABLE IF NOT EXISTS bucket (key TEXT PRIMARY KEY, data BYTEA, epoch REAL);')
+                conn.commit()
 
-        if USING_REDIS:
+        elif USING_REDIS:
             self.redis_client = redis.Redis(host=REDIS_HOST, port=REDIS_PORT, db=0)
             self.set = self.redis_set
             self.get = self.redis_get
@@ -150,7 +157,7 @@ class Client:
         if USING_REDIS:
             self.redis_client.connection.disconnect()
 
-    # @retry_connection
+    @retry_connection
     def set(self, key: str, data: bytes, timeout: float | None = 60 * 5.) -> None:
         """
         Set data for a given key on the server.
@@ -160,12 +167,13 @@ class Client:
             data (bytes): The data to store.
             timeout (float, optional): The timeout for the socket connection in seconds. Defaults to 60 * 5.
         """
-        with self.db.connect() as conn:
-            cur = conn.cursor()
-            cur.execute('INSERT INTO bucket (key, data, epoch) VALUES (%s, %s, %s) '
-                         'ON CONFLICT (key) DO UPDATE SET (data, epoch) = (EXCLUDED.data, EXCLUDED.epoch)', (key, data, time.time()))
-            conn.commit()
-        return
+        if USING_POSTGRES:
+            with self.db.connect() as conn:
+                cur = conn.cursor()
+                cur.execute('INSERT INTO bucket (key, data, epoch) VALUES (%s, %s, %s) '
+                             'ON CONFLICT (key) DO UPDATE SET (data, epoch) = (EXCLUDED.data, EXCLUDED.epoch)', (key, data, time.time()))
+                conn.commit()
+            return
         # raise RuntimeError('Not using redis.')
         with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
             s.setsockopt(socket.IPPROTO_TCP, socket.TCP_NODELAY, 1)
@@ -192,18 +200,14 @@ class Client:
         Returns:
             bytes or None: The retrieved data, or None if the key doesn't exist.
         """
-        # try:
-        with self.db.connect() as conn:
-            cur = conn.cursor()
-            cur.execute('SELECT data FROM bucket WHERE key = %s', (key,))
-            data = cur.fetchone()
-            if not data:
-                return None
-            return bytes(data[0])
-            data = cur.fetchone()[0]['data']
-            return data
-        # except: pass
-        return
+        if USING_POSTGRES:
+            with self.db.connect() as conn:
+                cur = conn.cursor()
+                cur.execute('SELECT data FROM bucket WHERE key = %s', (key,))
+                data = cur.fetchone()
+                if not data:
+                    return None
+                return bytes(data[0])
         # raise RuntimeError('Not using redis.')
         with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
             s.setsockopt(socket.IPPROTO_TCP, socket.TCP_NODELAY, 1)
@@ -231,11 +235,13 @@ class Client:
             key (str): The key to delete data for.
             timeout (float, optional): The timeout for the socket connection in seconds. Defaults to 60 * 5.
         """
-        with self.db.connect() as conn:
-            cur = conn.cursor()
-            cur.execute('DELETE FROM bucket WHERE key = %s', (key, ))
-            conn.commit()
-        return
+        if USING_POSTGRES:
+            with self.db.connect() as conn:
+                cur = conn.cursor()
+                cur.execute('DELETE FROM bucket WHERE key = %s', (key, ))
+                conn.commit()
+            return
+
         # raise RuntimeError('Not using redis.')
         with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
             s.setsockopt(socket.IPPROTO_TCP, socket.TCP_NODELAY, 1)
